+++ .todo.js
@@ -1,0 +1,2 @@
+// TODO Search adb exists in default folders before each command
+// TODO Check and execute adb commands in adb bin folder.