+++ src/Infraestructure/console/console-interface/index.ts
@@ -1,0 +1,9 @@
+import { execSync, ExecSyncOptions } from 'child_process'
+import { IConsoleInterface } from './iconsole-interface'
+
+export class ConsoleInterface implements IConsoleInterface {
+  execConsoleSync(command: string, options: ExecSyncOptions = null): Buffer {
+    console.log('Executed ', command)
+    return execSync(command, options)
+  }
+}