+++ src/adb-wrapper/index.ts
@@ -1,83 +1,78 @@
+import * as os from 'os'
 import adbCommands from './adb-commands'
-import { ConsoleChannel, consoleReturnAre } from '../console/console-channel'
+import {
+  ConsoleChannel,
+  consoleReturnAre
+} from '../Infraestructure/console/console-channel'
 import adbReturns from './adb-returns'
 import adbMessages from './adb-messages'
-import { NetHelpers } from '../net-helpers'
+import { NetHelpers } from '../Infraestructure/net-helpers'
 import { IPHelpers } from './ip-helpers'
 import { DeviceHelpers } from './device-helpers'
+import { ADBResolver } from '../adb-resolver'
+import { IConsoleInterface } from '../Infraestructure/console/console-interface/iconsole-interface'
+import { log } from 'util'
 
-const mapIPToDeviceName = (ipAddress: string): string => {
-  let extractedIP = IPHelpers.extractIPRegex(ipAddress)
-  try {
-    const nameOfDevice = DeviceHelpers.getDeviceModel(
-      this.consoleInstance,
-      extractedIP
-    )
-    return `${extractedIP} | ${nameOfDevice}`
-  } catch (e) {
-    return `${extractedIP} | NO DEVICE INFO`
+export class ADBConnection extends ConsoleChannel {
+  private resolverInstance: ADBResolver
+
+  constructor(ciInstance: IConsoleInterface) {
+    super(ciInstance)
+    this.resolverInstance = new ADBResolver(os.homedir(), os.type(), ciInstance)
   }
-}
-export class ADBChannel extends ConsoleChannel {
+
   /**
    *  connect to a given ip address
-   * @param ipAddress "192.168.1.100"
+   * @param ip "192.168.1.100"
    */
-  public ConnectToDevice(ipAddress: string): ADBResult {
+  public async ConnectToDevice(ip: string): Promise<string> {
     let finalResult = null
 
-    const deviceIP = IPHelpers.extractIPRegex(ipAddress)
-    const result = this.consoleInstance.execConsoleSync(
+    const deviceIP = IPHelpers.extractIPRegex(ip)
+    const result = await this.resolverInstance.sendADBCommand(
       adbCommands.CONNECT_IP_AND_PORT(deviceIP, '5555')
     )
-    const output = result.toString()
+    const resultString = result.toString()
     const deviceName = DeviceHelpers.getDeviceModel(
       this.consoleInstance,
       deviceIP
     )
 
-    if (consoleReturnAre(output, adbReturns.CONNECTED_TO())) {
-      finalResult = new ADBResult(
-        ADBResultState.ConnectedToDevice,
-        `Connected to: ${deviceName}`
-      )
+    if (consoleReturnAre(resultString, adbReturns.CONNECTED_TO())) {
+      finalResult = `Connected to: ${deviceName}`
     }
-    if (consoleReturnAre(output, adbReturns.ALLREADY_CONNECTED_TO())) {
+    if (consoleReturnAre(resultString, adbReturns.ALLREADY_CONNECTED_TO())) {
       throw new ADBInterfaceException(`Allready connected to: ${deviceName}`)
     }
-    if (consoleReturnAre(output, adbReturns.CONNECTION_REFUSED(deviceIP))) {
+    if (
+      consoleReturnAre(resultString, adbReturns.CONNECTION_REFUSED(deviceIP))
+    ) {
       throw new ADBInterfaceException(
         'Connection refused:\n Target machine actively refused connection.'
       )
     }
-    if (consoleReturnAre(output, adbReturns.MISSING_PORT(deviceIP))) {
+    if (consoleReturnAre(resultString, adbReturns.MISSING_PORT(deviceIP))) {
       throw new ADBInterfaceException('Port is missing fail to connect in ADB.')
     }
 
-    if (finalResult == null) {
-      throw new ADBInterfaceError('Some error ocurred during connection.')
+    if (finalResult === null) {
+      throw new ADBInterfaceException('ADB returned null value')
     }
     return finalResult
   }
-  public async ResetPorts(): Promise<ADBResult> {
+  public async ResetPorts(): Promise<string> {
     let finalResult = null
     try {
-      const result = this.consoleInstance.execConsoleSync(
+      const consoleReturn = await this.resolverInstance.sendADBCommand(
         adbCommands.RESET_PORTS()
       )
-      const output = result.toString()
+      const output = consoleReturn.toString()
       if (consoleReturnAre(output, adbReturns.RESTARTING_PORT())) {
-        finalResult = new ADBResult(
-          ADBResultState.DevicesInPortMode,
-          adbMessages.DEVICES_IN_TCP_MODE()
-        )
+        finalResult = adbMessages.DEVICES_IN_TCP_MODE()
       }
     } catch (e) {
       if (e.message.includes(adbReturns.NO_DEVICES_FOUND())) {
-        finalResult = new ADBResult(
-          ADBResultState.NoDevices,
-          adbMessages.NO_DEVICES_FOUND()
-        )
+        finalResult = adbMessages.NO_DEVICES_FOUND()
       } else {
         throw new ADBInterfaceException(e.message)
       }
@@ -88,18 +83,15 @@ export class ADBChannel extends ConsoleChannel {
     return finalResult
   }
 
-  public async DisconnectFromAllDevices(): Promise<ADBResult> {
+  public async DisconnectFromAllDevices(): Promise<string> {
     let finalResult = null
     try {
-      const result = this.consoleInstance.execConsoleSync(
+      const result = await this.resolverInstance.sendADBCommand(
         adbCommands.ADB_DISCONNECT_ALL()
       )
       const output = result.toString()
       if (consoleReturnAre(output, adbReturns.DISCONNECTED_EVERTHING())) {
-        finalResult = new ADBResult(
-          ADBResultState.DisconnectedEverthing,
-          'Disconnected from all devices'
-        )
+        finalResult = 'Disconnected from all devices'
       }
     } catch (e) {
       throw new ADBInterfaceError(e.toString())
@@ -109,10 +101,11 @@ export class ADBChannel extends ConsoleChannel {
     }
     return finalResult
   }
+
   public async FindConnectedDevices(): Promise<Array<string>> {
     let devicesArray = []
     try {
-      const result = this.consoleInstance.execConsoleSync(
+      const result = await this.resolverInstance.sendADBCommand(
         adbCommands.LIST_ADB_DEVICES()
       )
       const output = result.toString()
@@ -126,7 +119,19 @@ export class ADBChannel extends ConsoleChannel {
         }
 
         // try to get device name trought adb
-        ips = ips.map(mapIPToDeviceName)
+        ips = ips.map((ipAddress: string): string => {
+          let extractedIP = IPHelpers.extractIPRegex(ipAddress)
+          try {
+            const nameOfDevice = DeviceHelpers.getDeviceModel(
+              this.consoleInstance,
+              extractedIP
+            )
+            return `${extractedIP} | ${nameOfDevice}`
+          } catch (e) {
+            return `${extractedIP} | NO DEVICE INFO`
+          }
+        })
+
         return ips
       }
     } catch (e) {
@@ -138,19 +143,23 @@ export class ADBChannel extends ConsoleChannel {
     return devicesArray
   }
 
-  public async KillADBServer(): Promise<ADBResult> {
+  public async KillADBServer(): Promise<string> {
     let returned = null
     try {
-      const result = this.consoleInstance.execConsoleSync(
+      const result = await this.resolverInstance.sendADBCommand(
         adbCommands.ADB_KILL_SERVER()
       )
       if (result.toString() == adbReturns.ADB_KILLED_SUCCESS_RETURN()) {
-        returned = new ADBResult(ADBResultState.Success, 'ADB Server killed')
+        returned = 'ADB Server killed'
       } else {
-        throw Error('Internal error ocurred')
+        throw new ADBInterfaceError('ADB Server not killed')
       }
     } catch (e) {
-      throw new ADBInterfaceException(e.message)
+      if (e instanceof ADBInterfaceError) {
+        throw e
+      } else {
+        throw new ADBInterfaceException(e.message)
+      }
     }
     if (returned == null) {
       throw new ADBInterfaceError('Fail during ADB Kill')
@@ -159,39 +168,15 @@ export class ADBChannel extends ConsoleChannel {
   }
 }
 
-/**
- * Is an enum of adb possible results
- */
-export enum ADBResultState {
-  ConnectedToDevice,
-  NotFound,
-  NoDevices,
-  DevicesInPortMode,
-  AllreadyConnected,
-  DisconnectedEverthing,
-  Success
-}
-
-/**
- * Is an result returned by an adb connection
- */
-export class ADBResult {
-  state: ADBResultState
-  message: string
-  constructor(resultState: ADBResultState, message: string) {
-    this.state = resultState
-    this.message = message
-    return this
-  }
-}
-
 export class ADBInterfaceError extends Error {
-  AdbINterFaceError(message: string) {
-    this.message = adbMessages.ADB_INTERFACE_ERROR_DEFAULT()
+  constructor(message: string = adbMessages.ADB_INTERFACE_ERROR_DEFAULT()) {
+    super(message)
+    this.name = 'ADBInterfaceError'
   }
 }
 export class ADBInterfaceException extends Error {
-  ADBInterfaceException(message: string) {
-    this.message = message ?? adbMessages.ADB_INTERFACE_EXCEPTION_DEFAULT()
+  constructor(message: string = adbMessages.ADB_INTERFACE_EXCEPTION_DEFAULT()) {
+    super(message)
+    this.name = 'ADBInterfaceException'
   }
 }