+++ src/console/console-interface/iconsole-interface.ts
@@ -1,3 +1,0 @@
-export interface IConsoleInterface {
-  execConsoleSync(command: string): Buffer
-}