+++ src/console/console-interface/index.ts
@@ -1,8 +1,0 @@
-import { execSync } from 'child_process'
-import { IConsoleInterface } from './iconsole-interface'
-
-export class ConsoleInterface implements IConsoleInterface {
-  execConsoleSync(command: string): Buffer {
-    return execSync(command)
-  }
-}