+++ src/extension.ts
@@ -1,8 +1,14 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode'
-import { commandsRegistry } from './commands-registry'
+import { FirebaseController } from './controllers/firebase-controller'
+import { FirebaseManagerChannel } from './firebase-channel'
+import { ADBConnection } from './adb-wrapper'
+import { ADBCommandsController } from './controllers/adb-controller'
+import { ConsoleInterface } from './Infraestructure/console/console-interface/index'
 
+let firebaseController
+let adbCmdController
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
@@ -13,13 +19,14 @@ export function activate(context: vscode.ExtensionContext) {
   // Now provide the implementation of the command with registerCommand
   // The commandId parameter must match the command field in package.json
 
-  commandsRegistry(context).map(command => {
-    let subscription = vscode.commands.registerCommand(
-      command.name,
-      command.callback
-    )
-    context.subscriptions.push(subscription)
-  })
+  firebaseController = new FirebaseController(
+    context,
+    new FirebaseManagerChannel(new ConsoleInterface())
+  )
+  adbCmdController = new ADBCommandsController(
+    context,
+    new ADBConnection(new ConsoleInterface())
+  )
 }
 
 // this method is called when your extension is deactivated