+++ CHANGELOG.md
@@ -1,14 +1,27 @@
 # Changelog
 
+-   0.18.1
+
+    NEW - Refactor at codebase with `plugin controllers` startegy.
+    NEW - Changed Folder structure.
+
+-   0.18.0
+
+    NEW - Added test cases for `adb-resolver` use cases.
+
+    NEW - Improve how extension search `ADB` in enviroment adding `adb-resolver` strategy.
+
+    FIX - Enviroment variables is now accessed directly from path.
+
 -   0.17.2
 
-    Introduced fixes for issue #7 removing necessity to close "Connecting to X" message
+    FIX - for issue #7 removing necessity to close "Connecting to X" message
 
 -   0.17.1
 
-    Added fully ci/cd enviroment, now new commits in master will be released to production on extensions.
+    NEW - fully ci/cd enviroment, now new commits in master will be released to production on extensions.
 
--   Older versions
+-   Past
 
-    Changes forbiden
+    NEW - Forbiden Changes
     ...